[MISC] lot's of cleanup
