+++ src/lambda.cpp
@@ -19,8 +19,6 @@
 // lambda.cpp: Main File for Lambda
 // ==========================================================================
 
-#include "seqan2seqan3.hpp"
-
 #include <seqan3/argument_parser/all.hpp>
 
 #include "search.hpp"
