+++ src/shared_misc.hpp
@@ -32,41 +32,9 @@
 #include <seqan3/io/sequence_file/input.hpp>
 #include <seqan3/range/views/to_upper.hpp>
 
-// #include <seqan/basic.h>
-// #include <seqan/sequence.h>
-// #include <seqan/seq_io.h>
-// #include <seqan/index.h>
-// #include <seqan/align.h>
-// #include <seqan/blast.h>
-
-// using namespace seqan;
-
-// ============================================================================
-// Forwards
-// ============================================================================
-
-// ============================================================================
-// Metafunctions
-// ============================================================================
-
-// makes partial function specialization convenient
-template <bool condition>
-using MyEnableIf = typename std::enable_if<condition, int>::type;
-
 // ============================================================================
 // Functions for translation and retranslation
 // ============================================================================
-
-// template <typename TAlph>
-// inline std::basic_ostream<char> &
-// operator<<(std::basic_ostream<char> & out,
-//            const seqan::Iter<const seqan::String<seqan::SimpleType<unsigned char,TAlph>,
-//                                     seqan::Packed<> >,
-//                       seqan::Packed<> > it)
-// {
-//     out << *it;
-//     return out;
-// }
 
 template <typename TPos>
 inline bool
@@ -150,120 +118,6 @@ inline AlphabetEnum detectSeqFileAlphabet(std::string const & path)
 }
 
 // ----------------------------------------------------------------------------
-// Function readRecord(Fasta); an overload that truncates Ids at first Whitespace
-// ----------------------------------------------------------------------------
-
-// template <typename TSeqStringSet, typename TSpec, typename TRunnable>
-// inline void
-// _myReadRecordsImpl(TCDStringSet<seqan::String<char>> & meta,
-//                    TSeqStringSet & seq,
-//                    FormattedFile<Fastq, Input, TSpec> & file,
-//                    TRunnable && runnable)
-// {
-//     typedef typename SeqFileBuffer_<TSeqStringSet, TSpec>::Type TSeqBuffer;
-//
-//     TSeqBuffer seqBuffer;
-//
-//     // reuse the memory of context(file).buffer for seqBuffer (which has a different type but same sizeof(Alphabet))
-//     swapPtr(seqBuffer.data_begin, context(file).buffer[1].data_begin);
-//     swapPtr(seqBuffer.data_end, context(file).buffer[1].data_end);
-//     seqBuffer.data_capacity = context(file).buffer[1].data_capacity;
-//
-//     for (uint64_t count = 0; !atEnd(file); ++count) // count not used for abort condition
-//     {
-//         readRecord(context(file).buffer[0], seqBuffer, file);
-//
-//         // run whatever magic we are pushing in:
-//         runnable(context(file).buffer[0], count);
-//
-//         appendValue(meta, context(file).buffer[0]);
-//         appendValue(seq, seqBuffer);
-//     }
-//
-//     swapPtr(seqBuffer.data_begin, context(file).buffer[1].data_begin);
-//     swapPtr(seqBuffer.data_end, context(file).buffer[1].data_end);
-//     context(file).buffer[1].data_capacity = seqBuffer.data_capacity;
-//     seqBuffer.data_capacity = 0;
-// }
-
-// ----------------------------------------------------------------------------
-// Generic Sequence loading
-// ----------------------------------------------------------------------------
-
-// template <typename TSpec1,
-//           typename TSpec2,
-//           typename TFile,
-//           typename TRunnable>
-// void
-// myReadRecords(TCDStringSet<seqan::String<char, TSpec1>> & ids,
-//               TCDStringSet<seqan::String<seqan::Dna5, TSpec2>> & seqs,
-//               TFile                              & file,
-//               TRunnable                         && runnable)
-// {
-//     TCDStringSet<seqan::String<seqan::Iupac>> tmpSeqs; // all IUPAC nucleic acid characters are valid input
-//     try
-//     {
-//         _myReadRecordsImpl(ids, tmpSeqs, file, std::forward<TRunnable>(runnable));
-//     }
-//     catch(ParseError const & e)
-//     {
-//         std::string err;
-//         err += "\nParseError thrown: ";
-//         err += e.what();
-//         err += "\nMake sure that the file is standards compliant. If you get an unexpected character warning "
-//                "make sure you have set the right program parameter (-p), i.e. "
-//                "Lambda expected nucleic acid alphabet, maybe the file was protein?\n";
-//         throw std::runtime_error(err);
-//     }
-//
-//     seqs = tmpSeqs; // convert IUPAC alphabet to Dna5
-// }
-//
-// template <typename TSpec1,
-//           typename TSpec2,
-//           typename TFile,
-//           typename TRunnable>
-// void
-// myReadRecords(TCDStringSet<seqan::String<char, TSpec1>>       & ids,
-//               TCDStringSet<seqan::String<seqan::AminoAcid, TSpec2>>  & seqs,
-//               TFile                                    & file,
-//               TRunnable                               && runnable)
-// {
-//     try
-//     {
-//         _myReadRecordsImpl(ids, seqs, file, std::forward<TRunnable>(runnable));
-//     }
-//     catch(seqan::ParseError const & e)
-//     {
-//         std::string err;
-//         err += "\nParseError thrown: ";
-//         err += e.what();
-//         err += "\nMake sure that the file is standards compliant.\n";
-//         throw std::runtime_error(err);
-//     }
-//
-//     if (seqan::length(seqs) > 0)
-//     {
-//         // warn if sequences look like DNA
-//         if (seqan::CharString(seqan::String<seqan::Dna5>(seqan::CharString(seqs[0]))) == seqan::CharString(seqs[0]))
-//             std::cout << "\nWarning: The first query sequence looks like nucleic acid, but amino acid is expected.\n"
-//                          "           Make sure you have set the right program parameter (-p).\n";
-//     }
-// }
-//
-// template <typename TSpec1,
-//           typename TCharSpec,
-//           typename TSpec2,
-//           typename TFile>
-// void
-// myReadRecords(TCDStringSet<seqan::String<char, TSpec1>>       & ids,
-//               TCDStringSet<seqan::String<TCharSpec, TSpec2>>  & seqs,
-//               TFile                                    & file)
-// {
-//     myReadRecords(ids, seqs, file, [] (auto const &, uint64_t const) {});
-// }
-
-// ----------------------------------------------------------------------------
 // print if certain verbosity is set
 // ----------------------------------------------------------------------------
 
@@ -273,20 +127,6 @@ myPrintImpl(SharedOptions const & /**/,
             T const & first)
 {
     std::cout << first;
-}
-
-inline void
-myPrintImpl(SharedOptions const & options,
-            std::stringstream const & first)
-{
-    std::string str = first.str();
-//     std::cerr << "terminal cols: " << options.terminalCols
-//               << " str.size() " << str.size() << "\n";
-    if (options.isTerm && (str.size() >= (options.terminalCols -12)))
-        std::cout << str.substr(str.size()-options.terminalCols+12,
-                                options.terminalCols);
-    else
-        std::cout << str;
 }
 
 template <typename T, typename ... Args>
@@ -299,75 +139,15 @@ myPrintImpl(SharedOptions const & options,
     myPrintImpl(options, args...);
 }
 
-// template <typename ... Args>
-// inline void
-// myPrintImplThread(SharedOptions const & options,
-// //                   T const & first,
-//                   Args const & ... args)
-// {
-//     SEQAN_OMP_PRAGMA(critical(stdout))
-//     {
-// //                 std::cout << "\033[" << omp_get_thread_num() << "B";
-// //                 std::cout << "\033E";
-//         if (options.isTerm)
-//         {
-//             for (unsigned char i=0; i< omp_get_thread_num(); ++i)
-//                 std::cout << std::endl;
-//             std::cout << "\033[K";
-//         }
-//         std::cout << "Thread " << std::setw(3) << omp_get_thread_num() << "| ";
-//
-//         myPrintImpl(options, args...);
-//         std::cout << "\n" << std::flush;
-//         if (options.isTerm)
-//             std::cout << "\033[" << omp_get_thread_num()+1 << "A";
-//     }
-// }
-
 template <typename... Args>
 inline void
 myPrint(SharedOptions const & options, const int verbose, Args const &... args)
 {
     if (options.verbosity >= verbose)
     {
-//         #if defined(_OPENMP)
-//         if (omp_in_parallel())
-//             myPrintImplThread(options, args...);
-//         else
-//         #endif
-            myPrintImpl(options, args...);
-
+        myPrintImpl(options, args...);
         std::cout << std::flush;
     }
-}
-
-template <typename T>
-inline void
-appendToStatusImpl(std::stringstream & status,
-                   T const & first)
-{
-    status << first;
-}
-
-template <typename T, typename ... Args>
-inline void
-appendToStatusImpl(std::stringstream & status,
-                   T const & first,
-                   Args const & ... args)
-{
-    appendToStatusImpl(status, first);
-    appendToStatusImpl(status, args...);
-}
-
-template <typename... Args>
-inline void
-appendToStatus(std::stringstream & status,
-               SharedOptions const & options,
-               const int verbose,
-               Args const & ... args)
-{
-    if (options.verbosity >= verbose)
-        appendToStatusImpl(status, args...);
 }
 
 // ----------------------------------------------------------------------------
