+++ src/shared_options.hpp
@@ -26,229 +26,14 @@
 #include <thread>
 #include <unistd.h>
 
-#include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/alphabet/aminoacid/aa27.hpp>
-#include <seqan3/alphabet/aminoacid/aa10murphy.hpp>
-#include <seqan3/alphabet/aminoacid/aa10li.hpp>
-#include <seqan3/alphabet/aminoacid/translation_genetic_code.hpp>
 #include <seqan3/argument_parser/all.hpp>
 #include <seqan3/std/filesystem>
 
-// #include <seqan/basic.h>
-// #include <seqan/modifier.h>
-// #include <seqan/arg_parse.h>
-// #include <seqan/index.h>
-// #include <seqan/blast.h>
-// #include <seqan/misc/terminal.h>
-// #include <seqan/translation.h>
-// #include <seqan/reduced_aminoacid.h>
-
-// using namespace seqan;
-
-// ==========================================================================
-// Classes
-// ==========================================================================
-
-// --------------------------------------------------------------------------
-// Enum DbIndexType
-// --------------------------------------------------------------------------
-
-enum class DbIndexType : uint8_t
-{
-    FM_INDEX,
-    BI_FM_INDEX
-};
-
-inline std::string
-_indexEnumToName(DbIndexType const t)
-{
-    switch (t)
-    {
-        case DbIndexType::FM_INDEX:      return "fm_index";
-        case DbIndexType::BI_FM_INDEX:   return "bi_fm_index";
-    }
-
-    throw std::runtime_error("Error: unknown index type");
-    return "";
-}
-
-inline DbIndexType
-_indexNameToEnum(std::string const t)
-{
-    if (t == "bi_fm_index")
-        return DbIndexType::BI_FM_INDEX;
-    else if (t == "fm_index")
-        return DbIndexType::FM_INDEX;
-
-    throw std::runtime_error("Error: unknown index type");
-    return DbIndexType::FM_INDEX;
-}
-
-
-// --------------------------------------------------------------------------
-// Enum AlphabetEnum
-// --------------------------------------------------------------------------
-
-constexpr const char *
-_alphTypeToName(seqan3::dna4 const & /**/)
-{
-    return "dna4";
-}
-
-constexpr const char *
-_alphTypeToName(seqan3::dna5 const & /**/)
-{
-    return "dna5";
-}
-
-constexpr const char *
-_alphTypeToName(seqan3::aa27 const & /**/)
-{
-    return "aminoacid";
-}
-
-constexpr const char *
-_alphTypeToName(seqan3::aa10murphy const & /**/)
-{
-    return "murphy10";
-}
-
-constexpr const char *
-_alphTypeToName(seqan3::aa10li const & /**/)
-{
-    return "li10";
-}
-
-enum class AlphabetEnum : uint8_t
-{
-    UNDEFINED,
-    DNA4,
-    DNA5,
-    AMINO_ACID,
-    MURPHY10,
-    LI10,
-};
-
-inline std::string
-_alphabetEnumToName(AlphabetEnum const t)
-{
-    switch (t)
-    {
-        case AlphabetEnum::UNDEFINED:   return "UNDEFINED";
-        case AlphabetEnum::DNA4:        return _alphTypeToName(seqan3::dna4{});
-        case AlphabetEnum::DNA5:        return _alphTypeToName(seqan3::dna5{});
-        case AlphabetEnum::AMINO_ACID:  return _alphTypeToName(seqan3::aa27{});
-        case AlphabetEnum::MURPHY10:    return _alphTypeToName(seqan3::aa10murphy{});
-        case AlphabetEnum::LI10:        return _alphTypeToName(seqan3::aa10li{});
-    }
-
-    throw std::runtime_error("Error: unknown alphabet type");
-    return "";
-}
-
-inline AlphabetEnum
-_alphabetNameToEnum(std::string const t)
-{
-    if ((t == "UNDEFINED") || (t == "auto"))
-        return AlphabetEnum::UNDEFINED;
-    else if (t == _alphTypeToName(seqan3::dna4{}))
-        return AlphabetEnum::DNA4;
-    else if (t == _alphTypeToName(seqan3::dna5{}))
-        return AlphabetEnum::DNA5;
-    else if (t == _alphTypeToName(seqan3::aa27{}))
-        return AlphabetEnum::AMINO_ACID;
-    else if (t == _alphTypeToName(seqan3::aa10murphy{}))
-        return AlphabetEnum::MURPHY10;
-    else if (t == _alphTypeToName(seqan3::aa10li{}))
-        return AlphabetEnum::LI10;
-
-    throw std::runtime_error("Error: unknown alphabet type");
-    return AlphabetEnum::DNA4;
-}
-
-template <AlphabetEnum e>
-struct _alphabetEnumToType_;
-
-template <>
-struct _alphabetEnumToType_<AlphabetEnum::DNA4>
-{
-    using type = seqan3::dna4;
-};
-
-template <>
-struct _alphabetEnumToType_<AlphabetEnum::DNA5>
-{
-    using type = seqan3::dna5;
-};
-
-template <>
-struct _alphabetEnumToType_<AlphabetEnum::AMINO_ACID>
-{
-    using type = seqan3::aa27;
-};
-
-template <>
-struct _alphabetEnumToType_<AlphabetEnum::MURPHY10>
-{
-    using type = seqan3::aa10murphy;
-};
-
-template <>
-struct _alphabetEnumToType_<AlphabetEnum::LI10>
-{
-    using type = seqan3::aa10li;
-};
-
-template <AlphabetEnum e>
-using _alphabetEnumToType = typename _alphabetEnumToType_<e>::type;
-
-// inline uint64_t
-// _alphabetEnumToSize(AlphabetEnum const t)
-// {
-//     switch (t)
-//     {
-//         case AlphabetEnum::DNA4:        return sizeof(SizeTypePos_<Dna>);
-//         case AlphabetEnum::DNA5:        return sizeof(SizeTypePos_<Dna5>);
-//         case AlphabetEnum::AMINO_ACID:  return sizeof(SizeTypePos_<AminoAcid>);
-//         case AlphabetEnum::MURPHY10:    return sizeof(SizeTypePos_<ReducedAminoAcid<Murphy10>>);
-//     }
-//
-//     throw std::runtime_error("Error: unknown alphabet type");
-//     return 0;
-// }
-
-struct index_file_options
-{
-    uint64_t indexGeneration{0}; // bump this on incompatible changes
-
-    DbIndexType indexType{};
-
-    AlphabetEnum origAlph{};
-    AlphabetEnum transAlph{};
-    AlphabetEnum redAlph{};
-
-    seqan3::genetic_code geneticCode{};
-
-    //TODO reserve space here for more vars?
-
-    template <typename TArchive>
-    void serialize(TArchive & archive)
-    {
-        archive(cereal::make_nvp("generation", indexGeneration),
-                cereal::make_nvp("index type", indexType),
-                cereal::make_nvp("orig alph", origAlph),
-                cereal::make_nvp("trans alph", transAlph),
-                cereal::make_nvp("red alph", redAlph),
-                cereal::make_nvp("genetic code", geneticCode));
-    }
-};
+#include "shared_definitions.hpp"
 
 // --------------------------------------------------------------------------
 // Class SharedOptions
 // --------------------------------------------------------------------------
-
-// This struct stores the options from the command line.
 
 struct SharedOptions
 {
@@ -261,7 +46,6 @@ struct SharedOptions
 
     index_file_options indexFileOptions{};
 
-//     seqan::BlastProgram blastProgram   = seqan::BlastProgram::UNKNOWN;
     bool        nucleotide_mode = false;
     bool        need_to_translate = false;
 
@@ -283,14 +67,6 @@ struct SharedOptions
     }
 };
 
-
-// ==========================================================================
-// Functions
-// ==========================================================================
-
-// --------------------------------------------------------------------------
-// Function sharedSetup()
-// --------------------------------------------------------------------------
 
 inline void sharedSetup(seqan3::argument_parser & parser)
 {
